Update: 사용자 식별자 타입 변경 (Int -> Long)

+++ src/main/kotlin/kea/dpang/auth/entity/User.kt
@@ -20,7 +20,7 @@ data class User(
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    var userIdx: Int? = null,
+    var userIdx: Long? = null,
 
     var email: String? = null,
