+++ src/main/kotlin/kea/dpang/auth/service/UserService.kt
@@ -19,7 +19,7 @@ interface UserService {
      * 사용자의 이메일과 비밀번호를 입력받아 사용자를 검증합니다.
      * 검증에 성공한 경우, 사용자의 고유 식별자인 userIdx를 반환합니다.
      */
-    fun verifyUser(email: String, password: String): Int
+    fun verifyUser(email: String, password: String): Long
 
     /**
      * 비밀번호 재설정 요청 메서드.
